+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
-
 ## Added
 - [#19](https://github.com/practicalli/project-templates/issues/19) `practicalli/minimal` template
 - templates: link to Practicalli Clojure Project templates section for help after new project is created
+- [#22](https://github.com/practicalli/project-templates/pull/22) Service template with `:component` option for `:donut` and `:integrant`
+
 ### Changed
 - ci: update MegaLinter GitHub Action v7 in MegaLnter workflow
 - ci: set monthly schedule and cron examples for alternative schedules